+++ database/migrations/2026_05_25_070137_modify_total_jumlah_nullable_in_bahan_masuk.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up()
     {
         Schema::table('bahan_masuk', function (Blueprint $table) {
-            $table->integer('jumlah_total')->nullable()->change();
+            $table->decimal('jumlah_total', 15,2)->nullable()->change();
         });
     }
